pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sports-prediction-system"
version = "1.0.0"
description = "An advanced sports forecasting system for educational purposes"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Sports Prediction Team", email = "team@example.com"}
]
maintainers = [
    {name = "Sports Prediction Team", email = "team@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["sports", "prediction", "machine-learning", "forecasting", "analytics"]
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "lightgbm>=4.0.0",
    "xgboost>=2.0.0",
    "optuna>=3.4.0",
    "mlflow>=2.8.0",
    "shap>=0.42.0",
    "requests>=2.31.0",
    "aiohttp>=3.8.0",
    "pydantic>=2.4.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "plotly>=5.17.0",
    "dash>=2.14.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "jsonschema>=4.19.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pillow>=10.0.0",
    "reportlab>=4.0.0",
    "tqdm>=4.66.0",
    "joblib>=1.3.0",
    "python-slugify>=8.0.0",
    "rich>=13.6.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "bandit>=1.7.0",
]
deep-learning = [
    "torch>=2.1.0",
    "tensorflow>=2.14.0",
]
monitoring = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
]
bayesian = [
    "pymc>=5.7.0",
    "arviz>=0.16.0",
]

[project.urls]
Homepage = "https://github.com/username/sports-prediction-system"
Documentation = "https://github.com/username/sports-prediction-system/docs"
Repository = "https://github.com/username/sports-prediction-system.git"
"Bug Tracker" = "https://github.com/username/sports-prediction-system/issues"

[project.scripts]
sports-forecast = "app.cli:main"

[tool.setuptools]
packages = ["app", "features", "models", "dashboard"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.json", "*.csv", "*.sql"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
line-length = 88
target-version = "py39"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app", "features", "models", "dashboard"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    'if __name__ == .__main__.:',
    'class .*\\bProtocol\\):',
    '@(abc\\.)?abstractmethod',
]